requirements.txt
# Core
numpy>=1.24
scipy>=1.10
Pillow>=10.0
matplotlib>=3.7
pyyaml>=6.0

# Geospatial
rasterio>=1.3
shapely>=2.0
geopandas>=0.14
fiona>=1.9
pyproj>=3.6

# Deep Learning (training)
torch>=2.1
timm>=0.9
einops>=0.7
huggingface_hub>=0.20

# Image processing
opencv-python-headless>=4.8

# System monitoring (dashboard)
psutil>=5.9

# Data processing
pandas>=2.0

# Satellite data (needed for data fetching via DES)
openeo>=0.20
requests>=2.31

# Visualization helpers
svgpath2mpl>=0.2

# Notebook development (optional)
# jupyter>=1.0
# ipykernel>=6.0

# Shared contracts — 23-class schema, Analyzer base, KG JSON Schema,
# embedding config. Extracted from this repo in 2026-04 (rollout W3.6).
des-contracts @ git+https://github.com/TobiasEdman/des-contracts.git@v0.1.0

# HTTP API (W4.1). Optional in dev; required in the deployed Docker
# image. Pinned tight — fastapi 0.x has occasional breaking minor bumps.
fastapi>=0.115,<0.120
uvicorn[standard]>=0.32,<1
gunicorn>=22,<24
httpx>=0.27,<1   # used by /analyze callback hook + by tests

# SLA telemetry (W4.5). Optional — telemetry module is fail-open when
# this isn't installed; metrics endpoint still serves a placeholder.
prometheus-client>=0.20,<1

# Super-resolution showcase (imint/analyzers/sr/). Optional in dev —
# each wrapper raises a clear error if its package is missing, so the
# showcase still renders the models that are available. Pinned loose
# until the ESA OpenSR stack stabilises on PyPI.
# - opensr-utils:  production wrapper (sliding-window, edge-blend)
# - opensr-model:  LDSR-S2 latent diffusion
# - opensr-test:   benchmark (consistency / synthesis / hallucination)
opensr-utils; extra == "sr"
opensr-model; extra == "sr"
opensr-test;  extra == "sr"
